pyproject.toml
[project]
name = "langchain-agent-dev-tuto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.25.13",
    "deepagents>=0.4.9",
    "langchain>=1.2.11",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-qwq>=0.3.4",
    "langgraph-cli[inmem]>=0.4.15",
    "rich>=14.3.3",
    "volcengine>=1.0.216",
    "volcengine-python-sdk[ark]>=5.0.15",
    "aiohttp>=3.11.0",
    "psycopg[binary,pool]>=3.3.3",
    "langgraph>=1.1.2",
    "langgraph-checkpoint-postgres>=3.0.4",
    "jupyter>=1.1.1",
    "lab>=8.9",
    "tavily-python>=0.7.23",
    "markdownify>=1.2.2",
]

[tool.uv.workspace]
members = [
    "agent_server",
]
